Move variable name tweaking into the compiler
